pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "virtuosity-detector"
version = "0.1.0"
description = "Detection and analysis of vocal virtuosity in passerine birds using YOLOv8"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

keywords = ["bioacoustics", "bird song", "yolov8", "spectrogram", "ornithology"]

dependencies = [
    # Computer vision / deep learning
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # Audio / signal processing
    "librosa>=0.10.0",
    "soundfile>=0.12.0",

    # Data science
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",

    # Visualisation
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Configuration
    "pyyaml>=6.0",

    # Utilitaires
    "tqdm>=4.65.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "ipywidgets>=8.1.0",
]

[project.scripts]
virtuosity = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*", "*.egg-info"]
